requirements.txt
django==5.2.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0